viz/rlhf_module.py
import torch
import gymnasium as gym
from stable_baselines3 import PPO
from stable_baselines3.common.vec_env import DummyVecEnv
from stable_baselines3.common.callbacks import BaseCallback
from typing import Dict, Any
import numpy as np

class RLHFTrainer:
    def __init__(self, env_id="CartPole-v1"):  # Placeholder env; replace with custom AInception env
        self.env = gym.make(env_id)
        self.model = PPO("MlpPolicy", self.env, verbose=1, device="cuda" if torch.cuda.is_available() else "cpu")
        self.feedback_buffer = []  # Store human/AI feedback for RLHF

    def train_with_feedback(self, total_timesteps: int = 10000, feedback_data: list = None):
        """
        Train PPO model with RLHF-style feedback.
        feedback_data: List of (state, action, reward, feedback_score) tuples
        """
        if feedback_data:
            # Placeholder: Adjust rewards based on feedback
            for state, action, reward, feedback in feedback_data:
                adjusted_reward = reward + feedback  # Simple adjustment
                self.feedback_buffer.append((state, action, adjusted_reward))
                print(f"Adjusted reward: {adjusted_reward}")

        self.model.learn(total_timesteps=total_timesteps)
        return self.model

    def get_policy_action(self, obs: np.ndarray) -> tuple:
        """Get action from trained policy."""
        obs_tensor = np.array([obs])
        action, _states = self.model.predict(obs_tensor, deterministic=False)
        return action[0]

    def collect_feedback(self, state: np.ndarray, action: int, reward: float, human_score: float = 0.0) -> float:
        """Collect feedback for future training."""
        feedback_entry = (state, action, reward, human_score)
        self.feedback_buffer.append(feedback_entry)
        return human_score

# Example usage
if __name__ == "__main__":
    env = gym.make("CartPole-v1")  # Define env here
    trainer = RLHFTrainer()
    
    # Train with sample feedback
    feedback_data = [
        (np.array([0.1, 0.2]), 0, 1.0, 0.8),  # Good action
        (np.array([0.5, 0.6]), 1, -1.0, -0.5)   # Bad action
    ]
    model = trainer.train_with_feedback(1000, feedback_data)
    
    # Test policy
    obs, _ = env.reset()
    action = trainer.get_policy_action(obs)
    print("Policy Action:", action)
