requirements.txt
numpy==1.24.0
pyyaml==6.0
transformers==4.30.0
torch==2.1.0
PyQt6==6.5.0
diffusers==0.21.4
jax==0.4.20
jaxlib==0.4.20
accelerate==0.21.0
safetensors==0.3.3
huggingface-hub==0.16.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
stable-baselines3==2.0.0
torchvision==0.16.0
torchaudio==2.1.0
opencv-python==4.8.1.78
openai-whisper==20231117
torch_geometric==2.5.3
optuna==3.3.0
ray==2.6.3
lava==0.4.1
vtk==9.3.0
